Update to use new url for search and query.
Also updated query so that it correctly parses the results returned
and handles multiple ids correctly.

I've also removed the error check, which would never work as it
attempted to regex an array. In addition, PubMed simply returns an empty page
if there are no results, rather than an error message.